Create strcount.c
Given an Input string aaaabbccdee. Count the no of same consecutive character and give the output as a4b2c2d1e2. Assume: Maximum Length of the string is 20. 
Based on the above example, write a program that accept the input string through STDIN and writes the transformed string to STDOUT.

Other than the transformed string, no other character / string / message should be written to STDOUT.